refactor: extract shared computeHostId into a single location
