+++ clients/shared/Network/GatewayHTTPClient.swift
@@ -1,8 +1,4 @@
 import Foundation
-#if os(macOS)
-import CryptoKit
-import IOKit
-#endif
 
 /// Authenticated HTTP client for gateway and platform proxy requests.
 ///
@@ -517,29 +513,9 @@ public enum GatewayHTTPClient {
 
     #if os(macOS)
     /// Compute a stable device ID from the IOPlatformUUID.
+    /// Delegates to the shared `HostIdComputer` implementation.
     private static func computeMacOSDeviceId() -> String {
-        let platformUUID = getMacOSPlatformUUID() ?? UUID().uuidString
-        let salt = "vellum-assistant-host-id"
-        let input = Data((platformUUID + salt).utf8)
-        let hash = SHA256.hash(data: input)
-        return hash.compactMap { String(format: "%02x", $0) }.joined()
-    }
-
-    /// Read the IOPlatformUUID from the IORegistry (macOS hardware identifier).
-    private static func getMacOSPlatformUUID() -> String? {
-        let service = IOServiceGetMatchingService(
-            kIOMainPortDefault,
-            IOServiceMatching("IOPlatformExpertDevice")
-        )
-        guard service != 0 else { return nil }
-        defer { IOObjectRelease(service) }
-
-        let key = kIOPlatformUUIDKey as CFString
-        guard let uuid = IORegistryEntryCreateCFProperty(service, key, kCFAllocatorDefault, 0)?
-            .takeRetainedValue() as? String else {
-            return nil
-        }
-        return uuid
+        return HostIdComputer.computeHostId()
     }
     #endif
 }
